requirements.txt
celery
Flask
Flask-Mail
python-dotenv